Revert "seo and new Hero component"
