+++ package.json
@@ -24,7 +24,6 @@
     "react-router-dom": "^7.1.3",
     "react-scripts": "^5.0.1",
     "sass": "^1.83.4",
-    "sass-loader": "^16.0.4",
     "sitemap": "^8.0.0",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
