+++ src/components/About/About.scss
@@ -1,4 +1,4 @@
-@use '../../assets/styles/shared' as *;
+@import '../../assets/styles/shared';
 
 #about {
   text-align: left;
