+++ src/components/Contact/Contact.scss
@@ -1,5 +1,5 @@
 
-@use '../../assets/styles/shared' as *;
+@import '../../assets/styles/shared';
 
 .contact-section {
   position: relative;
