+++ src/components/Navbar/Navbar.scss
@@ -1,4 +1,4 @@
-@use '../../assets/styles/shared' as *;
+@import '../../assets/styles/shared';
 
 //Navbar common properties
 $transparent-bg: transparent !important;
